chore: resolved review comment.
